Setup Git basics and add project README
Configured Git file interactions (both file ignores and file attributes)
in addition to adding a Git commit message template which will be
enabled on a subsequent commit. Also created a project README file using
Markdown so that it will be "pretty" when displayed on the GitHub page
of the project.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>