Setup bare-bones Maven build
Create a very simple Maven POM which can use the nar-maven-plugin to
create NAR archives containing the JNI headers. The nar-maven-plugin is
able to build on a wide variety of operating systems and development
environments, but this project will be using the GCC project's g++ to
compile and link as necessary. The Maven build is also configured via
the .mvn directory, which contains both build extensions and execution
options for the project build.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>