Enable and configure Travis-CI builds
Add the Travis-CI configuration file to the root of the project to
enable continuous integration builds to begin running for the project.

Note: due to the nature of this project requiring an environment
containing both a Java Development Kit and the GCC project's compilers
the current configuration is very likely to require additional
modifications to reach a passing CI state. Once the CI build is working
an additional commit indicating the success will be created.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>