Attempt to fix Travis-CI configuration
Modify .travis.yml file to use apt-get packages add-on instead of a
matrix include approach involving multiple language tag declarations. It
is not possible to know if this commit will fix the CI build until after
it is pushed upstream. As a result this commit may turn out to not solve
the problem, in which case additional fix commits will be forthcoming.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>