Replace buildnumber-maven-plugin
Replace the buildnumber-maven-plugin with the git-commit-id-plugin to
generate build unique information. The new configuration adds a
buildinfo file to the root of the project's JAR artifact containing
useful and unique info about the state of the git repository upon the
generation/building of that artifact. A small modification to the
travis-ci configuration was also made to add caching of the Maven M2
repository between CI builds.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>